requirements.txt
flet
transformers
torch
sentencepiece
Mastodon.py
praw
discord.py
python-dotenv
tqdm
psycopg2-binary
sqlalchemy
facebook-sdk
fpdf2